Updated Version 2.1 with simulated DoorOpen contact
Updated Version 2.1 with simulated DoorOpen contact
Works perfect from HK side. Have some sync problem when running from KNX side because of this current issue: 
In this version I have a problem that for some reason my master If lines (Line: 101, 107, 113, 133) regarding the HK Characteristic CurrentDoorState is never get true. It looks like I don't get the right value when I read it in the If()
Any idea what is the reason?